requirements.txt
matplotlib==3.3.3
glfw==2.0.0
pyrr==0.10.3
pandas==1.2.0
numpy==1.19.3
tensorflow==2.1.0
Pillow==8.1.0
progressbar2==3.53.1
PyOpenGL==3.1.5
scikit_learn==0.24.0
